requirements.txt
fastapi
pydantic
uvicorn
numpy
nltk
PyPDF2
scikit-learn
sentence-transformers
torch
transformers
huggingface-hub
python-dotenv
shutil
glob2
